requirements.txt
boto3==1.4.7
python-twitter==3.3
